+++ ProdManager/helpers/config.py
@@ -1,0 +1,3 @@
+
+def boolean_param(string):
+  return string.lower() in ('true', '1', 't')