+++ ProdManager/routes/incident/views.py
@@ -15,12 +15,12 @@ from ProdManager.helpers.resource import (
 from ProdManager.helpers.date import current_date
 from ProdManager.helpers.json import json_defaults
 from ProdManager.helpers.form import strip_input
+from ProdManager.helpers.notification import send_notification
 
-from ProdManager.models.Incident import Incident, IncidentSeverity, IncidentStatus
-from ProdManager.models.Scope import Scope
-from ProdManager.models.Service import Service
-from ProdManager.models.Event import EventType
-from ProdManager.models.IncidentEvent import IncidentEvent
+from ProdManager.models import (
+  Incident, IncidentSeverity, IncidentStatus, Scope,
+  Service, EventType, IncidentEvent,
+)
 
 from .forms import IncidentCreateForm, IncidentUpdateForm, IncidentCommentForm, IncidentDeleteForm
 
@@ -87,6 +87,20 @@ def create():
     ))
   except Exception as error:
     current_app.logger.error(f"Unable to create event during Incident creation : {error}")
+
+  try:
+    notif_title = f"[{incident.severity.name}][{incident.status.name}] {incident.name} - New Incident"
+    if incident.external_reference:
+      notif_title = f"[{incident.external_reference}]{notif_title}"
+
+    send_notification(
+      notif_title,
+      render_template("notification/incident.html",
+        incident=incident,
+      )
+    )
+  except Exception as error:
+    current_app.logger.error(f"Unable to send notification during Incident creation : {error}")
 
   return redirect(url_for('incident.show', resource_id=incident.id), 302)
 
@@ -181,6 +195,20 @@ def update(resource_id):
       ))
     except Exception as error:
       current_app.logger.error(f"Unable to create event during Incident update : {error}")
+
+    try:
+      notif_title = f"[{incident.severity.name}][{incident.status.name}] {incident.name} - Updated Incident"
+      if incident.external_reference:
+        notif_title = f"[{incident.external_reference}]{notif_title}"
+
+      send_notification(
+        notif_title,
+        render_template("notification/incident.html",
+          incident=incident,
+        )
+      )
+    except Exception as error:
+      current_app.logger.error(f"Unable to send notification during Incident update : {error}")
 
   return redirect(url_for('incident.show', resource_id=incident.id), 302)