+++ ProdManager/routes/maintenance/views.py
@@ -14,12 +14,12 @@ from ProdManager.helpers.resource import (
 from ProdManager.helpers.date import current_date
 from ProdManager.helpers.json import json_defaults
 from ProdManager.helpers.form import strip_input
+from ProdManager.helpers.notification import send_notification
 
-from ProdManager.models.Maintenance import Maintenance, MaintenanceStatus
-from ProdManager.models.Scope import Scope
-from ProdManager.models.Service import Service, ServiceStatus
-from ProdManager.models.Event import EventType
-from ProdManager.models.MaintenanceEvent import MaintenanceEvent
+from ProdManager.models import (
+  Maintenance, MaintenanceStatus, Scope, Service,
+  ServiceStatus, EventType, MaintenanceEvent,
+)
 
 from .forms import (
   MaintenanceCreateForm, MaintenanceUpdateForm, MaintenanceCommentForm,
@@ -91,6 +91,20 @@ def create():
     ))
   except Exception as error:
     current_app.logger.error(f"Unable to create event during Maintenance creation : {error}")
+
+  try:
+    notif_title = f"[{maintenance.status.name}] {maintenance.name} - Scheduled Maintenance"
+    if maintenance.external_reference:
+      notif_title = f"[{maintenance.external_reference}]{notif_title}"
+
+    send_notification(
+      notif_title,
+      render_template("notification/maintenance.html",
+        maintenance=maintenance,
+      )
+    )
+  except Exception as error:
+    current_app.logger.error(f"Unable to send notification during Maintenance creation : {error}")
 
   return redirect(url_for('maintenance.show', resource_id=maintenance.id), 302)
 
@@ -180,6 +194,19 @@ def update(resource_id):
       ))
     except Exception as error:
       current_app.logger.error(f"Unable to create event during Maintenance update : {error}")
+
+    try:
+      notif_title = f"[{maintenance.status.name}] {maintenance.name} - Updated Maintenance"
+      if maintenance.external_reference:
+        notif_title = f"[{maintenance.external_reference}]{notif_title}"
+      send_notification(
+        notif_title,
+        render_template("notification/maintenance.html",
+          maintenance=maintenance,
+        )
+      )
+    except Exception as error:
+      current_app.logger.error(f"Unable to send notification during Maintenance update : {error}")
 
   return redirect(url_for('maintenance.show', resource_id=maintenance.id), 302)